fix : suspense 사용 빌드 에러 해결
